+++ src/config/cloudinaryConfig.js
@@ -1,0 +1,23 @@
+import { config, uploader } from 'cloudinary'
+
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const cloudinaryConfig = (req, res, next) => {
+
+config({
+
+cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+
+api_key: process.env.CLOUDINARY_API_KEY,
+
+api_secret: process.env.CLOUDINARY_SECRET,
+
+});
+
+next();
+
+}
+
+export { cloudinaryConfig, uploader };
